requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.31.0
openai
pytest
pytest-asyncio
pytest-mock
pytest-cov # para cobertura de pruebas unitarias
loguru
ruff # para linting y formateo